shell script updated

+++ deployment.sh
@@ -21,8 +21,15 @@ cd "$temp_dir/$DOCKERFILE_PATH"
 docker build -t "$DOCKERHUB_USERNAME/$DOCKER_IMAGE" .
 
 echo "-------------Docker Run-------------"
-docker container rm -f "$DOCKER_CONT_NAME"
+if docker ps -q --filter "name=$DOCKER_CONT_NAME" >/dev/null; then
+    echo "Stopping and removing the running container..."
+    docker container stop "$DOCKER_CONT_NAME"
+    docker container rm -f "$DOCKER_CONT_NAME"
+else
+    echo "No running container with the name $DOCKER_CONT_NAME found."
+fi
 docker run -d -p 80:80 --name "$DOCKER_CONT_NAME" "$DOCKERHUB_USERNAME/$DOCKER_IMAGE:latest"
 
 echo "-------------CleanUp the Workspace-------------"
 rm -rf $temp_dir
+
